refactor: update mock import paths
