+++ src/components/profile/__tests__/ProfileEditor.test.tsx
@@ -5,7 +5,7 @@ import { render, screen, waitFor, act } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { useConnectedAccountsStore } from '@/lib/stores/connected-accounts.store';
 import { OAuthProvider } from '@/types/oauth';
-import { createConnectedAccountsStoreMock } from '@/tests/mocks/connected-accounts.store.mock';
+import { createConnectedAccountsStoreMock } from '@/services/user/__tests__/mocks/connected-accounts.store.mock';
 import { vi, Mock } from 'vitest';
 import type { UserEvent } from '@testing-library/user-event';
